# Backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Real Estate",
    "description": "A Real Estate project with multiple roles",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "admin@realestate.com"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
...
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register API",
        "description": "Add a new user",
        "requestBody": {
          "description": "Create a new pet in the store",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "country_code": {
                    "type": "string"
                  },
                  "contact": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "about": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "USER",
                      "AGENT"
                    ]
                  },
                  "device_token": {
                    "type": "string"
                  },
                  "device_type": {
                    "type": "string"
                  },
                  "app_version": {
                    "type": "string"
                  },
                  "device_model": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/verifyOtp": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register API",
        "description": "Verify Otp",
        "requestBody": {
          "description": "Verifying otp",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "otp": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Login API",
        "description": "Login a registered user",
        "requestBody": {
          "description": "Login API",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "login_type": {
                    "type": "string",
                    "enum": [
                      "NORMAL"
                    ]
                  },
                  "device_token": {
                    "type": "string"
                  },
                  "device_type": {
                    "type": "string"
                  },
                  "app_version": {
                    "type": "string"
                  },
                  "device_model": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/updateLicense": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "update license API",
        "description": "update a license",
        "parameters": [
          {
            "name": "x_token",
            "in": "header",
            "description": "This is description",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "license_number": {
                    "type": "string",
                    "default": ""
                  },
                  "city": {
                    "type": "string",
                    "default": ""
                  },
                  "zipcode": {
                    "type": "string",
                    "default": ""
                  },
                  "address": {
                    "type": "string",
                    "default": ""
                  },
                  "latitude": {
                    "type": "number",
                    "default": ""
                  },
                  "longitude": {
                    "type": "number",
                    "default": ""
                  },
                  "company_details": {
                    "type": {
                      "name": {
                        "type": "string",
                        "default": ""
                      },
                      "website": {
                        "type": "string",
                        "default": ""
                      },
                      "contact": {
                        "type": "string",
                        "default": ""
                      },
                      "address": {
                        "type": "string",
                        "default": ""
                      },
                      "latitude": {
                        "type": "number",
                        "default": ""
                      },
                      "longitude": {
                        "type": "number",
                        "default": ""
                      },
                      "city": {
                        "type": "string",
                        "default": ""
                      },
                      "zipcode": {
                        "type": "string",
                        "default": ""
                      }
                    }
                  }
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {}
                }
              }
            },
            "400": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "/user/socialLogin": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Social Login API",
        "description": "Add a new user",
        "requestBody": {
          "description": "Social Login API",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "USER",
                      "AGENT"
                    ]
                  },
                  "login_type": {
                    "type": "string",
                    "enum": [
                      "GOOGLE",
                      "FACEBOOK",
                      "NORMAL"
                    ]
                  },
                  "device_token": {
                    "type": "string"
                  },
                  "device_type": {
                    "type": "string"
                  },
                  "app_version": {
                    "type": "string"
                  },
                  "device_model": {
                    "type": "string"
                  },
                  "social_id": {
                    "type": "string",
                    "required": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/recoverAccount": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Recover account API",
        "description": "Recover Account",
        "requestBody": {
          "description": "Recover account API",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/verifyChangePassword": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "verify and recover account API",
        "description": "verify and recover Account",
        "requestBody": {
          "description": "verify and recover account API",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "otp": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}